Move tests to test/netflix/*
