+++ tests/netflix/test_cassandra.py
@@ -1,3 +1,5 @@
+import pytest
+
 from service_capacity_modeling.capacity_planner import planner
 from service_capacity_modeling.interface import AccessConsistency
 from service_capacity_modeling.interface import AccessPattern
@@ -16,6 +18,9 @@ from service_capacity_modeling.interface import FixedInterval
 from service_capacity_modeling.interface import GlobalConsistency
 from service_capacity_modeling.interface import Interval
 from service_capacity_modeling.interface import QueryPattern
+from service_capacity_modeling.models.org.netflix.cassandra import (
+    NflxCassandraCapacityModel,
+)
 
 small_but_high_qps = CapacityDesires(
     service_tier=1,
@@ -449,3 +454,45 @@ def test_preserve_memory():
 
     lr_clusters = cap_plan[0].candidate_clusters.zonal[0]
     assert lr_clusters.instance.ram_gib == 128
+
+
+@pytest.mark.parametrize(
+    "tier, extra_model_arguments, expected_result",
+    [
+        # Non-critical tier, no required_cluster_size
+        (2, {}, None),
+        # Non-critical tier, required_cluster_size provided
+        (2, {"required_cluster_size": 5}, 5),
+        # Critical tier, required_cluster_size >= CRITICAL_TIER_MIN_CLUSTER_SIZE
+        (0, {"required_cluster_size": 3}, 3),
+        (0, {"required_cluster_size": 2}, 2),
+        # Critical tier, no required_cluster_size
+        (0, {}, None),
+    ],
+)
+def test_get_required_cluster_size_valid(tier, extra_model_arguments, expected_result):
+    result = NflxCassandraCapacityModel.get_required_cluster_size(
+        tier, extra_model_arguments
+    )
+    assert result == expected_result
+
+
+@pytest.mark.parametrize(
+    "tier, extra_model_arguments",
+    [
+        # Critical tier(s), required_cluster_size < CRITICAL_TIER_MIN_CLUSTER_SIZE
+        (
+            1,
+            {"required_cluster_size": 1},
+        ),
+        (
+            0,
+            {"required_cluster_size": 1},
+        ),
+    ],
+)
+def test_get_required_cluster_size_exceptions(tier, extra_model_arguments):
+    with pytest.raises(ValueError):
+        NflxCassandraCapacityModel.get_required_cluster_size(
+            tier, extra_model_arguments
+        )
